move to client command
